Refactored: Extract to Configuration::Base from Settings.
